vllm plugin: rename nemotron_v3 to salm, collapse to single model class with backend composition
Address Piotr's review (PR #15520): the nemotron_v3 folder name no longer matches the scope (the plugin handles both standard transformer and hybrid Mamba+MoE backbones), and the _NeMoSpeechLMBase + two-class inheritance pattern duplicates wiring across backbones. Rename the package to salm and replace the inheritance structure with composition: a single NeMoSpeechLMForConditionalGeneration class delegates LLM-specific work (architecture name, weight rename, optional LoRA merge, mamba state passthroughs) to a TransformerBackend or HybridBackend selected by make_backend(config).

Single-class registration is feasible because vLLM's runtime ModelConfig.is_hybrid property uses text_config.layer_types as an escape hatch (the granite-4.0-micro path): we declare IsHybrid on the model class for the NemotronH backbone, and config.py populates layer_types=['attention']*N for transformer backbones so vLLM treats them as attention-only at runtime. There is no runtime isinstance(model, IsHybrid) check anywhere in vLLM that bypasses the property, so this collapses cleanly.

salm/ now splits into:
  config.py      NeMoSpeechLMConfig + the layer_types shim
  multimodal.py  audio helpers and vLLM processor/info/dummy-inputs trio
  backends.py    _BaseBackend, TransformerBackend, HybridBackend, make_backend
  model.py       single NeMoSpeechLMForConditionalGeneration class
  __init__.py    register() with one architecture name

examples/speechlm2/to_hf.py emits the unified architecture name. Tests are updated for the new import paths and add coverage for layer_types shim wiring and make_backend dispatch (45 unit tests pass).

Made-with: Cursor
