+++ pyproject.toml
@@ -67,7 +67,7 @@ py-modules = ["nemo"]
 llm = "nemo.collections.llm"
 
 [project.entry-points."vllm.general_plugins"]
-nemo_speechlm = "nemo.collections.speechlm2.vllm.nemotron_v3:register"
+nemo_speechlm = "nemo.collections.speechlm2.vllm.salm:register"
 
 [project.urls]
 Download = "https://github.com/NVIDIA/NeMo/releases"
